requirements.txt
gunicorn==20.1.0
PyYAML==5.4.1
WebOb==1.8.7